% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crime_plot.R
\name{CrimePlot}
\alias{CrimePlot}
\title{plot total offence according to specific area and type}
\usage{
CrimePlot(crime_data, offence_description, suburbs)
}
\arguments{
\item{crime_data}{A data.table object with the following columns:
"date" (POSIXct), "suburb" (chr), "postcode" (chr), "offence_level_1" (chr),
"offence_level_2" (chr), "offence_level_3" (chr), "offence_count" (num).}

\item{offence_description}{A character string of one description found in Offence level 3.}

\item{suburbs}{A two-element character vector. Each element is the name (UPPERCASE)
of an SA suburb.}
}
\value{
A ggplot object showing the correlation in offence count between the two input suburbs.
}
\description{
\code{CrimePlot} < it will accept any of the offence_descriptions found in Offence Level 3
and will accept a 2-element vector of suburbs, then plot the total offence of the specificed
description in the two suburbs in each month>
}
\examples{
<one or two examples showing how to use the function>
crime_data = crime[[1]]
offence_description = "OFFENCES AGAINST PROPERTY"
suburbs = c("ABERFOYLE PARK", "ADELAIDE")
CrimePlot(crime_data, offence_description, suburbs)
}
